+++ client/components/LandingPage.jsx
@@ -1,14 +1,14 @@
 import React, { Component, Fragment } from 'react'
 
 class LandingPage extends Component {
-    render(){
-        return(
+    render() {
+        return (
             <Fragment>
-            <div className='header'>
-            <h2>? Rocket Launch : Code Launch</h2>
-            {/* start coding here */}
+                <div className='header'>
+                    <h2>? Rocket Launch : Code Launch</h2>
+                    {/* start coding here */}
 
-            </div>
+                </div>
             </Fragment>
         )
     }
