+++ server/public/css/main.css
@@ -7,9 +7,6 @@
 
 #app {
   min-width: 100vw;
-}
-
-.rocket {
   min-height: 100vh;
   background-color: lightseagreen;
   display: flex;
@@ -18,15 +15,18 @@
 
 .rocket-panel {
   display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
 }
 
+
+
 .rocket-wrapper {
   height: 75%;
   width: 50%;
-}
+  min-width: 200px;
+
+}
